+++ docs/guides/extend/behavior_group.py
@@ -1,0 +1,7 @@
+from navground import core
+
+
+class MyBehaviorGroup(core.BehaviorGroup):
+
+    # MUST override
+    def compute_cmds(self, time_step: float) -> list[core.Twist2]: ...